+++ src/Resources/contao/dca/tl_vereinsregister.php
@@ -110,9 +110,16 @@ $GLOBALS['TL_DCA']['tl_vereinsregister'] = array
 			'toggle' => array
 			(
 				'label'               => &$GLOBALS['TL_LANG']['tl_vereinsregister']['toggle'],
-				'icon'                => 'visible.gif',
-				'attributes'          => 'onclick="Backend.getScrollOffset();return AjaxRequest.toggleVisibility(this,%s)"',
-				'button_callback'     => array('tl_vereinsregister', 'toggleIcon')
+				'attributes'           => 'onclick="Backend.getScrollOffset()"',
+				'haste_ajax_operation' => array
+				(
+					'field'            => 'published',
+					'options'          => array
+					(
+						array('value' => '', 'icon' => 'invisible.svg'),
+						array('value' => '1', 'icon' => 'visible.svg'),
+					),
+				),
 			),
 			'show' => array
 			(
@@ -133,7 +140,7 @@ $GLOBALS['TL_DCA']['tl_vereinsregister'] = array
 	// Subpalettes
 	'subpalettes' => array
 	(
-		'addImage'                    => 'singleSRC,fullsize,size,floating,overwriteMeta',
+		'addImage'                    => 'singleSRC,size,floating,fullsize,overwriteMeta',
 		'overwriteMeta'               => 'alt,imageTitle,imageUrl,caption',
 		'addBefore'                   => 'beforeClubs',
 		'addSeparation'               => 'separationClubs',
@@ -515,13 +522,6 @@ $GLOBALS['TL_DCA']['tl_vereinsregister'] = array
 			},
 			'sql'                     => "varchar(64) NOT NULL default ''"
 		),
-		'fullsize' => array
-		(
-			'label'                   => &$GLOBALS['TL_LANG']['tl_vereinsregister']['fullsize'],
-			'inputType'               => 'checkbox',
-			'eval'                    => array('tl_class'=>'w50'),
-			'sql'                     => array('type' => 'boolean', 'default' => false)
-		),
 		'floating' => array
 		(
 			'label'                   => &$GLOBALS['TL_LANG']['tl_vereinsregister']['floating'],
@@ -531,12 +531,19 @@ $GLOBALS['TL_DCA']['tl_vereinsregister'] = array
 			'reference'               => &$GLOBALS['TL_LANG']['MSC'],
 			'sql'                     => "varchar(12) NOT NULL default 'above'"
 		),
+		'fullsize' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_vereinsregister']['fullsize'],
+			'inputType'               => 'checkbox',
+			'eval'                    => array('tl_class'=>'w50', 'type' => 'boolean'),
+			'sql'                     => "char(1) NOT NULL default ''"
+		),
 		'overwriteMeta' => array
 		(
 			'label'                   => &$GLOBALS['TL_LANG']['tl_vereinsregister']['overwriteMeta'],
 			'inputType'               => 'checkbox',
-			'eval'                    => array('submitOnChange'=>true, 'tl_class'=>'w50 clr'),
-			'sql'                     => array('type' => 'boolean', 'default' => false)
+			'eval'                    => array('submitOnChange'=>true, 'tl_class'=>'w50 clr', 'type' => 'boolean'),
+			'sql'                     => "char(1) NOT NULL default ''"
 		),
 		'alt' => array
 		(
@@ -644,74 +651,6 @@ class tl_vereinsregister extends Backend
 		else $temp .= ' <span style="color:#b3b3b3;padding-left:3px">['.\Schachbulle\ContaoHelperBundle\Classes\Helper::getDate($arrRow['foundingDate']).' - '.\Schachbulle\ContaoHelperBundle\Classes\Helper::getDate($arrRow['resolutionDate']).']</span>';
 		$temp .= '</div>';
 		return $temp;
-	}
-
-	/**
-	 * Return the "toggle visibility" button
-	 * @param array
-	 * @param string
-	 * @param string
-	 * @param string
-	 * @param string
-	 * @param string
-	 * @return string
-	 */
-	public function toggleIcon($row, $href, $label, $title, $icon, $attributes)
-	{
-		$this->import('BackendUser', 'User');
-	
-		if (strlen($this->Input->get('tid')))
-		{
-			$this->toggleVisibility($this->Input->get('tid'), ($this->Input->get('state') == 0));
-			$this->redirect($this->getReferer());
-		}
-	
-		// Check permissions AFTER checking the tid, so hacking attempts are logged
-		if (!$this->User->isAdmin && !$this->User->hasAccess('tl_vereinsregister::published', 'alexf'))
-		{
-			return '';
-		}
-	
-		$href .= '&amp;id='.$this->Input->get('id').'&amp;tid='.$row['id'].'&amp;state='.$row[''];
-	
-		if (!$row['published'])
-		{
-			$icon = 'invisible.gif';
-		}
-	
-		return '<a href="'.$this->addToUrl($href).'" title="'.specialchars($title).'"'.$attributes.'>'.$this->generateImage($icon, $label).'</a> ';
-	}
-
-	/**
-	 * Disable/enable a user group
-	 * @param integer
-	 * @param boolean
-	 */
-	public function toggleVisibility($intId, $blnPublished)
-	{
-		// Check permissions to publish
-		if (!$this->User->isAdmin && !$this->User->hasAccess('tl_vereinsregister::published', 'alexf'))
-		{
-			$this->log('Not enough permissions to show/hide club record ID "'.$intId.'"', 'tl_vereinsregister toggleVisibility', TL_ERROR);
-			$this->redirect('contao/main.php?act=error');
-		}
-	
-		$this->createInitialVersion('tl_vereinsregister', $intId);
-	
-		// Trigger the save_callback
-		if (is_array($GLOBALS['TL_DCA']['tl_vereinsregister']['fields']['published']['save_callback']))
-		{
-			foreach ($GLOBALS['TL_DCA']['tl_vereinsregister']['fields']['published']['save_callback'] as $callback)
-			{
-				$this->import($callback[0]);
-				$blnPublished = $this->$callback[0]->$callback[1]($blnPublished, $this);
-			}
-		}
-	
-		// Update the database
-		$this->Database->prepare("UPDATE tl_vereinsregister SET tstamp=". time() .", published='" . ($blnPublished ? '' : '1') . "' WHERE id=?")
-			 ->execute($intId);
-		$this->createNewVersion('tl_vereinsregister', $intId);
 	}
 
 	public function getVereinsliste(DataContainer $dc)