+++ src/Resources/contao/dca/tl_vereinsregister_chronik.php
@@ -91,9 +91,16 @@ $GLOBALS['TL_DCA']['tl_vereinsregister_chronik'] = array
 			'toggle' => array
 			(
 				'label'               => &$GLOBALS['TL_LANG']['tl_vereinsregister_chronik']['toggle'],
-				'icon'                => 'visible.gif',
-				'attributes'          => 'onclick="Backend.getScrollOffset();return AjaxRequest.toggleVisibility(this,%s)"',
-				'button_callback'     => array('tl_vereinsregister_chronik', 'toggleIcon')
+				'attributes'           => 'onclick="Backend.getScrollOffset()"',
+				'haste_ajax_operation' => array
+				(
+					'field'            => 'published',
+					'options'          => array
+					(
+						array('value' => '', 'icon' => 'invisible.svg'),
+						array('value' => '1', 'icon' => 'visible.svg'),
+					),
+				),
 			),
 			'show' => array
 			(
@@ -107,15 +114,15 @@ $GLOBALS['TL_DCA']['tl_vereinsregister_chronik'] = array
 	// Palettes
 	'palettes' => array
 	(
-		'__selector__'                => array('addImage', 'protected'),
-		'default'                     => '{chronik_legend},date,headline,text;{source_legend:hide},source,url;{image_legend},addImage;{protected_legend:hide},protected;{expert_legend:hide},guests,cssID,space;{invisible_legend:hide},invisible,start,stop'
+		'__selector__'                => array('addImage', 'overwriteMeta'),
+		'default'                     => '{chronik_legend},date,headline,text;{source_legend:hide},source,url;{image_legend},addImage;{publish_legend:hide},published'
 	),
 
 	// Subpalettes
 	'subpalettes' => array
 	(
-		'addImage'                    => 'singleSRC,alt,title,size,imagemargin,imageUrl,fullsize,caption,floating',
-		'protected'                   => 'groups'
+		'addImage'                    => 'singleSRC,size,floating,fullsize,overwriteMeta',
+		'overwriteMeta'               => 'alt,imageTitle,imageUrl,caption',
 	),
 
 	// Fields
@@ -127,7 +134,7 @@ $GLOBALS['TL_DCA']['tl_vereinsregister_chronik'] = array
 		),
 		'pid' => array
 		(
-			'foreignKey'              => 'tl_vereinsregister.alias',
+			'foreignKey'              => 'tl_vereinsregister.name',
 			'sql'                     => "int(10) unsigned NOT NULL default '0'",
 			'relation'                => array('type'=>'belongsTo', 'load'=>'eager')
 		),
@@ -198,14 +205,51 @@ $GLOBALS['TL_DCA']['tl_vereinsregister_chronik'] = array
 		'singleSRC' => array
 		(
 			'label'                   => &$GLOBALS['TL_LANG']['tl_vereinsregister_chronik']['singleSRC'],
-			'exclude'                 => true,
 			'inputType'               => 'fileTree',
-			'eval'                    => array('filesOnly'=>true, 'fieldType'=>'radio', 'mandatory'=>true, 'tl_class'=>'clr'),
-			'sql'                     => "binary(16) NULL",
-			'save_callback' => array
+			'eval'                    => array
 			(
-				array('tl_vereinsregister_chronik', 'storeFileMetaInformation')
-			)
+				'fieldType'           => 'radio', 
+				'filesOnly'           => true, 
+				'extensions'          => '%contao.image.valid_extensions%', 
+				'mandatory'           => true,
+				'tl_class'            => 'clr'
+			),
+			'sql'                     => "binary(16) NULL"
+		),
+		'size' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['MSC']['imgSize'],
+			'inputType'               => 'imageSize',
+			'reference'               => &$GLOBALS['TL_LANG']['MSC'],
+			'eval'                    => array('rgxp'=>'natural', 'includeBlankOption'=>true, 'nospace'=>true, 'helpwizard'=>true, 'tl_class'=>'w50 clr'),
+			'options_callback' => static function ()
+			{
+				return Contao\System::getContainer()->get('contao.image.image_sizes')->getOptionsForUser(Contao\BackendUser::getInstance());
+			},
+			'sql'                     => "varchar(64) NOT NULL default ''"
+		),
+		'floating' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_vereinsregister_chronik']['floating'],
+			'inputType'               => 'radioTable',
+			'options'                 => array('above', 'left', 'right', 'below'),
+			'eval'                    => array('cols'=>4, 'tl_class'=>'w50'),
+			'reference'               => &$GLOBALS['TL_LANG']['MSC'],
+			'sql'                     => "varchar(12) NOT NULL default 'above'"
+		),
+		'fullsize' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_vereinsregister_chronik']['fullsize'],
+			'inputType'               => 'checkbox',
+			'eval'                    => array('tl_class'=>'w50', 'type' => 'boolean'),
+			'sql'                     => "char(1) NOT NULL default ''"
+		),
+		'overwriteMeta' => array
+		(
+			'label'                   => &$GLOBALS['TL_LANG']['tl_vereinsregister_chronik']['overwriteMeta'],
+			'inputType'               => 'checkbox',
+			'eval'                    => array('submitOnChange'=>true, 'tl_class'=>'w50 clr', 'type' => 'boolean'),
+			'sql'                     => "char(1) NOT NULL default ''"
 		),
 		'alt' => array
 		(
@@ -216,33 +260,14 @@ $GLOBALS['TL_DCA']['tl_vereinsregister_chronik'] = array
 			'eval'                    => array('maxlength'=>255, 'tl_class'=>'w50'),
 			'sql'                     => "varchar(255) NOT NULL default ''"
 		),
-		'title' => array
+		'imageTitle' => array
 		(
-			'label'                   => &$GLOBALS['TL_LANG']['tl_vereinsregister_chronik']['title'],
+			'label'                   => &$GLOBALS['TL_LANG']['tl_vereinsregister_chronik']['imageTitle'],
 			'exclude'                 => true,
 			'search'                  => true,
 			'inputType'               => 'text',
 			'eval'                    => array('maxlength'=>255, 'tl_class'=>'w50'),
 			'sql'                     => "varchar(255) NOT NULL default ''"
-		),
-		'size' => array
-		(
-			'label'                   => &$GLOBALS['TL_LANG']['tl_vereinsregister_chronik']['size'],
-			'exclude'                 => true,
-			'inputType'               => 'imageSize',
-			'options'                 => $GLOBALS['TL_CROP'],
-			'reference'               => &$GLOBALS['TL_LANG']['MSC'],
-			'eval'                    => array('rgxp'=>'digit', 'nospace'=>true, 'helpwizard'=>true, 'tl_class'=>'w50'),
-			'sql'                     => "varchar(64) NOT NULL default ''"
-		),
-		'imagemargin' => array
-		(
-			'label'                   => &$GLOBALS['TL_LANG']['tl_vereinsregister_chronik']['imagemargin'],
-			'exclude'                 => true,
-			'inputType'               => 'trbl',
-			'options'                 => array('px', '%', 'em', 'rem', 'ex', 'pt', 'pc', 'in', 'cm', 'mm'),
-			'eval'                    => array('includeBlankOption'=>true, 'tl_class'=>'w50'),
-			'sql'                     => "varchar(128) NOT NULL default ''"
 		),
 		'imageUrl' => array
 		(
@@ -250,20 +275,8 @@ $GLOBALS['TL_DCA']['tl_vereinsregister_chronik'] = array
 			'exclude'                 => true,
 			'search'                  => true,
 			'inputType'               => 'text',
-			'eval'                    => array('rgxp'=>'url', 'decodeEntities'=>true, 'maxlength'=>255, 'tl_class'=>'w50 wizard'),
-			'wizard' => array
-			(
-				array('\Schachbulle\ContaoVereinsregisterBundle\Classes\Vereinsregister', 'pagePicker')
-			),
+			'eval'                    => array('rgxp'=>'url', 'decodeEntities'=>true, 'maxlength'=>255, 'dcaPicker'=>true, 'addWizardClass'=>false, 'tl_class'=>'w50'),
 			'sql'                     => "varchar(255) NOT NULL default ''"
-		),
-		'fullsize' => array
-		(
-			'label'                   => &$GLOBALS['TL_LANG']['tl_vereinsregister_chronik']['fullsize'],
-			'exclude'                 => true,
-			'inputType'               => 'checkbox',
-			'eval'                    => array('tl_class'=>'w50 m12'),
-			'sql'                     => "char(1) NOT NULL default ''"
 		),
 		'caption' => array
 		(
@@ -273,17 +286,6 @@ $GLOBALS['TL_DCA']['tl_vereinsregister_chronik'] = array
 			'inputType'               => 'text',
 			'eval'                    => array('maxlength'=>255, 'allowHtml'=>true, 'tl_class'=>'w50'),
 			'sql'                     => "varchar(255) NOT NULL default ''"
-		),
-		'floating' => array
-		(
-			'label'                   => &$GLOBALS['TL_LANG']['tl_vereinsregister_chronik']['floating'],
-			'default'                 => 'above',
-			'exclude'                 => true,
-			'inputType'               => 'radioTable',
-			'options'                 => array('above', 'left', 'right', 'below'),
-			'eval'                    => array('cols'=>4, 'tl_class'=>'w50'),
-			'reference'               => &$GLOBALS['TL_LANG']['MSC'],
-			'sql'                     => "varchar(32) NOT NULL default ''"
 		),
 		'source' => array
 		(
@@ -303,73 +305,14 @@ $GLOBALS['TL_DCA']['tl_vereinsregister_chronik'] = array
 			'eval'                    => array('mandatory'=>false, 'rgxp'=>'url', 'decodeEntities'=>true, 'maxlength'=>255, 'tl_class'=>'w50'),
 			'sql'                     => "varchar(255) NOT NULL default ''"
 		),
-		'protected' => array
+		'published' => array
 		(
-			'label'                   => &$GLOBALS['TL_LANG']['tl_vereinsregister_chronik']['protected'],
+			'label'                   => &$GLOBALS['TL_LANG']['tl_vereinsregister_chronik']['published'],
 			'exclude'                 => true,
 			'filter'                  => true,
-			'inputType'               => 'checkbox',
-			'eval'                    => array('submitOnChange'=>true),
-			'sql'                     => "char(1) NOT NULL default ''"
-		),
-		'groups' => array
-		(
-			'label'                   => &$GLOBALS['TL_LANG']['tl_vereinsregister_chronik']['groups'],
-			'exclude'                 => true,
-			'inputType'               => 'checkbox',
-			'foreignKey'              => 'tl_member_group.name',
-			'eval'                    => array('mandatory'=>true, 'multiple'=>true),
-			'sql'                     => "blob NULL",
-			'relation'                => array('type'=>'hasMany', 'load'=>'lazy')
-		),
-		'guests' => array
-		(
-			'label'                   => &$GLOBALS['TL_LANG']['tl_vereinsregister_chronik']['guests'],
-			'exclude'                 => true,
-			'filter'                  => true,
-			'inputType'               => 'checkbox',
-			'eval'                    => array('tl_class'=>'w50'),
-			'sql'                     => "char(1) NOT NULL default ''"
-		),
-		'cssID' => array
-		(
-			'label'                   => &$GLOBALS['TL_LANG']['tl_vereinsregister_chronik']['cssID'],
-			'exclude'                 => true,
-			'inputType'               => 'text',
-			'eval'                    => array('multiple'=>true, 'size'=>2, 'tl_class'=>'w50 clr'),
-			'sql'                     => "varchar(255) NOT NULL default ''"
-		),
-		'space' => array
-		(
-			'label'                   => &$GLOBALS['TL_LANG']['tl_vereinsregister_chronik']['space'],
-			'exclude'                 => true,
-			'inputType'               => 'text',
-			'eval'                    => array('multiple'=>true, 'size'=>2, 'rgxp'=>'digit', 'nospace'=>true, 'tl_class'=>'w50'),
-			'sql'                     => "varchar(64) NOT NULL default ''"
-		),
-		'invisible' => array
-		(
-			'label'                   => &$GLOBALS['TL_LANG']['tl_vereinsregister_chronik']['invisible'],
-			'exclude'                 => true,
-			'filter'                  => true,
+			'default'                 => 1,
 			'inputType'               => 'checkbox',
 			'sql'                     => "char(1) NOT NULL default ''"
-		),
-		'start' => array
-		(
-			'exclude'                 => true,
-			'label'                   => &$GLOBALS['TL_LANG']['tl_vereinsregister_chronik']['start'],
-			'inputType'               => 'text',
-			'eval'                    => array('rgxp'=>'datim', 'datepicker'=>true, 'tl_class'=>'w50 wizard'),
-			'sql'                     => "varchar(10) NOT NULL default ''"
-		),
-		'stop' => array
-		(
-			'exclude'                 => true,
-			'label'                   => &$GLOBALS['TL_LANG']['tl_vereinsregister_chronik']['stop'],
-			'inputType'               => 'text',
-			'eval'                    => array('rgxp'=>'datim', 'datepicker'=>true, 'tl_class'=>'w50 wizard'),
-			'sql'                     => "varchar(10) NOT NULL default ''"
 		),
 	)
 );
@@ -399,9 +342,9 @@ class tl_vereinsregister_chronik extends Backend
 	{
 
 		// Status der Sichtbarkeit für das CSS feststellen
-		$key = $arrRow['invisible'] ? 'unpublished' : 'published';
+		$key = $arrRow['published'] ? 'published' : 'published';
 
-		$temp = '<div class="cte_type ' . $key . '">Ereignis' . ($arrRow['protected'] ? ' (' . $GLOBALS['TL_LANG']['MSC']['protected'] . ')' : ($arrRow['guests'] ? ' (' . $GLOBALS['TL_LANG']['MSC']['guests'] . ')' : '')) . '</div>';
+		$temp = '<div class="cte_type ' . $key . '">Eintrag</div>';
 
 		$temp .= '<div class="limit_height' . (!$GLOBALS['TL_CONFIG']['doNotCollapse'] ? ' h64' : '') . ' block">';
 		$temp .= '<div class="tl_gray">';
@@ -410,152 +353,9 @@ class tl_vereinsregister_chronik extends Backend
 		return $temp.'</div></div>';
 	}
 
-	/**
-	 * Pre-fill the "alt" and "caption" fields with the file meta data
-	 * @param mixed
-	 * @param \DataContainer
-	 * @return mixed
-	 */
-	public function storeFileMetaInformation($varValue, DataContainer $dc)
-	{
-		if ($dc->activeRecord->singleSRC == $varValue)
-		{
-			return $varValue;
-		}
-
-		$objFile = \FilesModel::findByUuid($varValue);
-
-		//if ($objFile !== null)
-		//{
-		//	$arrMeta = deserialize($objFile->meta);
-        //
-		//	if (!empty($arrMeta))
-		//	{
-		//		$objPage = $this->Database->prepare("SELECT * FROM tl_page WHERE id=(SELECT pid FROM " . ($dc->activeRecord->ptable ?: 'tl_article') . " WHERE id=?)")
-		//								  ->execute($dc->activeRecord->pid);
-        //
-		//		if ($objPage->numRows)
-		//		{
-		//			$objModel = new PageModel();
-		//			$objModel->setRow($objPage->row());
-		//			$objModel->loadDetails();
-        //
-		//			// Convert the language to a locale (see #5678)
-		//			$strLanguage = str_replace('-', '_', $objModel->rootLanguage);
-        //
-		//			if (isset($arrMeta[$strLanguage]))
-		//			{
-		//				\Input::setPost('alt', $arrMeta[$strLanguage]['title']);
-		//				\Input::setPost('caption', $arrMeta[$strLanguage]['caption']);
-		//			}
-		//		}
-		//	}
-		//}
-
-		return $varValue;
-	}
-
-	/**
-	 * Return the "toggle visibility" button
-	 * @param array
-	 * @param string
-	 * @param string
-	 * @param string
-	 * @param string
-	 * @param string
-	 * @return string
-	 */
-	public function toggleIcon($row, $href, $label, $title, $icon, $attributes)
-	{
-		if (strlen(Input::get('tid')))
-		{
-			$this->toggleVisibility(Input::get('tid'), (Input::get('state') == 1));
-			$this->redirect($this->getReferer());
-		}
-
-		// Check permissions AFTER checking the tid, so hacking attempts are logged
-		if (!$this->User->isAdmin && !$this->User->hasAccess('tl_vereinsregister_chronik::invisible', 'alexf'))
-		{
-			return '';
-		}
-
-		$href .= '&amp;id='.Input::get('id').'&amp;tid='.$row['id'].'&amp;state='.$row['invisible'];
-
-		if ($row['invisible'])
-		{
-			$icon = 'invisible.gif';
-		}
-
-		return '<a href="'.$this->addToUrl($href).'" title="'.specialchars($title).'"'.$attributes.'>'.Image::getHtml($icon, $label).'</a> ';
-	}
-
-
-	/**
-	 * Toggle the visibility of an element
-	 * @param integer
-	 * @param boolean
-	 */
-	public function toggleVisibility($intId, $blnVisible)
-	{
-		// Check permissions to edit
-		Input::setGet('id', $intId);
-		Input::setGet('act', 'toggle');
-
-		// The onload_callbacks vary depending on the dynamic parent table (see #4894)
-		if (is_array($GLOBALS['TL_DCA']['tl_vereinsregister_chronik']['config']['onload_callback']))
-		{
-			foreach ($GLOBALS['TL_DCA']['tl_vereinsregister_chronik']['config']['onload_callback'] as $callback)
-			{
-				if (is_array($callback))
-				{
-					$this->import($callback[0]);
-					$this->$callback[0]->$callback[1]($this);
-				}
-				elseif (is_callable($callback))
-				{
-					$callback($this);
-				}
-			}
-		}
-
-		// Check permissions to publish
-		if (!$this->User->isAdmin && !$this->User->hasAccess('tl_vereinsregister_chronik::invisible', 'alexf'))
-		{
-			$this->log('Not enough permissions to show/hide content element ID "'.$intId.'"', __METHOD__, TL_ERROR);
-			$this->redirect('contao/main.php?act=error');
-		}
-
-		$objVersions = new Versions('tl_vereinsregister_chronik', $intId);
-		$objVersions->initialize();
-
-		// Trigger the save_callback
-		if (is_array($GLOBALS['TL_DCA']['tl_vereinsregister_chronik']['fields']['invisible']['save_callback']))
-		{
-			foreach ($GLOBALS['TL_DCA']['tl_vereinsregister_chronik']['fields']['invisible']['save_callback'] as $callback)
-			{
-				if (is_array($callback))
-				{
-					$this->import($callback[0]);
-					$blnVisible = $this->$callback[0]->$callback[1]($blnVisible, $this);
-				}
-				elseif (is_callable($callback))
-				{
-					$blnVisible = $callback($blnVisible, $this);
-				}
-			}
-		}
-
-		// Update the database
-		$this->Database->prepare("UPDATE tl_vereinsregister_chronik SET tstamp=". time() .", invisible='" . ($blnVisible ? '' : 1) . "' WHERE id=?")
-					   ->execute($intId);
-
-		$objVersions->create();
-		$this->log('A new version of record "tl_vereinsregister_chronik.id='.$intId.'" has been created'.$this->getParentEntries('tl_vereinsregister_chronik', $intId), __METHOD__, TL_GENERAL);
-	}
-
 	public function groupFormat($group, $sortingMode, $firstOrderBy, $row, $dc)
 	{
-		return \Schachbulle\ContaoVereinsregisterBundle\Classes\Vereinsregister::getDateString($group);
+		return \Schachbulle\ContaoHelperBundle\Classes\Helper::getDate($group);
 
 	}
 }